test-requirements.txt
odoo-addon-rma_lot @ git+https://github.com/OCA/rma.git@refs/pull/431/head#subdirectory=rma_lot
odoo-addon-stock_picking_return_lot @ git+https://github.com/OCA/stock-logistics-workflow.git@refs/pull/1745/head#subdirectory=stock_picking_return_lot
